fixed css background issue
