+++ src/Components/CSS/footer.module.scss
@@ -18,7 +18,6 @@ $iconColor: white;
         display: flex;
         justify-content: space-evenly;
         align-items: center;
-        padding-left: 45px;
 
         .camera {
             color: $iconColor;
