+++ src/Components/CSS/login.module.scss
@@ -1,25 +1,33 @@
-*{
-    text-align: center;
+.Login {
+    height: 100vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    // text-align: center;
     /* background-image: url("/src/img/Component_4.png"); */
     background-size: 100%;
     background-position-x: -35px ;
     background: url("/src/img/Component_4.png") no-repeat center fixed;
 }
-h1{
+
+.Login > h1 {
     margin: 50px;
-    margin-top: 160px;    
 }
-.divarl{
+
+.divarl {
     margin: 30px;
 }
-.input{
+
+.input {
     height: 30px;
     border-radius: 5px;
     text-align: start;
     letter-spacing: 0.5pt; 
     background: white;  
 }
-.btnl{
+
+.btnl {
     background: white;
     color: black;
     width: 100px;
