+++ src/Components/CSS/register.module.scss
@@ -1,5 +1,11 @@
 .divr{
-    text-align: center;
+    height: 100vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+
+    // text-align: center;
     /* background-image: url("/src/img/Component_4.png"); */
     background-size: 130%;
     background-position-x: -35px ;
