+++ src/Components/Searchbar.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Style from './CSS/home.module.scss'
+import Style from './CSS/Home.module.scss'
 import TestImage from '../img/TestImage.jpg'
 
 function Searchbar() {
