+++ src/Pages/Register.jsx
@@ -1,36 +1,36 @@
 import React from 'react'
-import '../Components/CSS/register.css'
+import styles from '../Components/CSS/register.module.scss'
 import Navbar from '../Components/Navbar'
 
 function Register() {
   return (
-    <div className="divr">
-
+    <div>
        <Navbar />
-         <br />
+       <div className={styles.divr}>
          <h1>Register</h1>
                             
               <input type="username"
                      placeholder="Username..."
-                     className="inputl"
+                     className={styles.inputl}
               />              
               <br />        
               <input type="password"
                      placeholder="Password"
-                     className="inputl"
+                     className={styles.inputl}
               />                  
                <br />        
 
               <input type="email"
                      placeholder="Email"
-                     className="inputl"
+                     className={styles.inputl}
               />              
               <br />           
               <button type="submit"
-                    className="btn">Create</button>
+                    className={styles.btn}>Create</button>
 
                 <br />
               <small>Already have an account?<a href="#">Login here</a></small>             
+        </div>
     </div>
   )
 }
